+++ AntWars/Game.cs
@@ -25,7 +25,7 @@ namespace AntWars
 
         public Game(Config config)
         {
-            int startMoney = config.StartMoney;
+            double startMoney = config.StartMoney;
             Player1 = new Player(new AILoader(config.Player1AIPath), startMoney);
             Player2 = new Player(new AILoader(config.Player2AIPath), startMoney);
             initAI(Player1);
